refactor: reorganize storage-base-sql module
